// stack-over-clone/src/app/auth/types/authState.interface.ts
import {BackendErrorsInterface} from './../../shared/types/backendErrors.interface'
import {CurrentUserInterface} from 'src/app/shared/types/currentUser.interface'
export interface AuthStateInterface {
  isSubmitting: boolean
  currentUser: CurrentUserInterface | null
  isLoggedIn: boolean
  validationErrors: BackendErrorsInterface | null
  isLoading: boolean
}